commit 356f81882eee931e3cc62e8b1ca76c6f5c647109
  Author: iclsrc <iclsrc@intel.com>
  Date:   Wed May 18 20:23:13 2022 -0700

      Resolve of merge (build) of resolve of merge (conflict) 9d6bc3664a3a [OpenMP][Offload][L0] Enable sycl wrapper for interop (#49.. 'xmain' -> 'xmain-web' (2 commits) (#5113) (#5114)

      commit 88c1a0db553e2b9e9a57c901599f2303d5c2862d
        Author: iclsrc <iclsrc@intel.com>
        Date:   Wed May 18 20:14:50 2022 -0700

            Resolve of merge (conflict) 9d6bc3664a3a [OpenMP][Offload][L0] Enable sycl wrapper for interop (#49.. 'xmain' -> 'xmain-web' (2 commits) (#5113)

            * Resolve of merge (conflict) 9d6bc3664a3a [OpenMP][Offload][L0] Enable sycl wrapper for interop (#49.. 'xmain' -> 'xmain-web' (2 commits)

              CONFLICT (content): Merge conflict in openmp/libomptarget/plugins/level0/CMakeLists.txt

              commit 9d6bc3664a3aed9e37568dc736adb948f9db8b45
              Author: Hansang Bae <hansang.bae@intel.com>
              Date:   Wed May 18 18:41:21 2022 -0500

                  [OpenMP][Offload][L0] Enable sycl wrapper for interop (#4976)

                  Changes from the original PR.
                  - Moved intel_win_dlfcn.h to correct location
                  - Added missing wrapper API check after loading wrapper
                  - Enabled Windows build
                  - Updated copyright year in omptarget.rc

            * Resolve conflict

+++ openmp/libomptarget/include/device.h
@@ -468,7 +468,7 @@ struct DeviceTy {
   int32_t isSupportedDevice(void *DeviceType);
 #if INTEL_CUSTOMIZATION
   __tgt_interop *createInterop(int32_t InteropContext, int32_t NumPrefers,
-                               intptr_t *PreferIDs);
+                               int32_t *PreferIDs);
   int32_t releaseInterop(__tgt_interop *Interop);
   int32_t useInterop(__tgt_interop *Interop);
   int32_t getNumInteropProperties(void);
